refactor: extract self-contained helper types into focused files (move-only)
Batched, near-zero-risk relocations of already-clean types that were living
in oversized files, with zero behaviour change. Each type is moved verbatim
to a focused file and its (single) usage site updated; no public API changes.

- game_screen/models/game_list_models.dart: GameSaveDetectionData,
  LocalizedDescriptionData and their two background-task helpers, out of
  my_games_list.dart (they were unreferenced elsewhere in that file).
- utils/no_glow_scroll_behavior.dart: _NoGlowScrollBehavior ->
  public NoGlowScrollBehavior, out of secondary_screen.dart.
- widgets/system_emulator_settings_dialog/models/emulator_list_item.dart:
  the 5 EmulatorListItem data classes, out of system_emulator_settings_dialog.dart.
- utils/semaphore.dart: a single public Semaphore, replacing the two
  identical private _Semaphore copies in screenscraper_service.dart and
  optimized_md5_utils.dart (dedup).

Verified: flutter analyze — No issues found (project-wide); new files pass
dart format; flutter test 198/198 pass. Diffs are pure cut/paste.
